+++ addOns/authhelper/src/test/java/org/zaproxy/addon/authhelper/AuthUtilsUnitTest.java
@@ -33,6 +33,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
 
 import fi.iki.elonen.NanoHTTPD;
 import fi.iki.elonen.NanoHTTPD.IHTTPSession;
@@ -1052,6 +1053,8 @@ class AuthUtilsUnitTest extends TestUtils {
                             new String[] {"-headless"},
                             new String[] {"remote.active-protocols=1"},
                             Map.of("webSocketUrl", true)));
+
+            mockMessages(new ExtensionAuthhelper());
         }
 
         @BeforeEach
@@ -1194,6 +1197,63 @@ class AuthUtilsUnitTest extends TestUtils {
         }
 
         @TestTemplate
+        void shouldReturnPasswordFieldWithPasswordInName(WebDriver wd) {
+            // Given
+            pageContent =
+                    () ->
+                            """
+                                <input type="text" name="IsPasswordField" />
+                            """;
+            wd.get(url);
+            List<WebElement> inputElements = AuthUtils.getInputElements(wd, false);
+
+            // When
+            WebElement field = AuthUtils.getPasswordField(inputElements);
+
+            // Then
+            assertThat(field, is(notNullValue()));
+        }
+
+        @TestTemplate
+        void shouldReturnPasswordFieldWithPasswordInId(WebDriver wd) {
+            // Given
+            pageContent =
+                    () ->
+                            """
+                                <input type="text" id="IsPasswordField" />
+                            """;
+            wd.get(url);
+            List<WebElement> inputElements = AuthUtils.getInputElements(wd, false);
+
+            // When
+            WebElement field = AuthUtils.getPasswordField(inputElements);
+
+            // Then
+            assertThat(field, is(notNullValue()));
+        }
+
+        @TestTemplate
+        void shouldReturnPasswordFieldByTypeOverIdAndName(WebDriver wd) {
+            // Given
+            pageContent =
+                    () ->
+                            """
+                                <input type="text" name="IsPasswordField" />
+                                <input type="text" id="IsPasswordField" />
+                                <input type="password" />
+                            """;
+            wd.get(url);
+            List<WebElement> inputElements = AuthUtils.getInputElements(wd, false);
+
+            // When
+            WebElement field = AuthUtils.getPasswordField(inputElements);
+
+            // Then
+            assertThat(field, is(notNullValue()));
+            assertThat(field.getDomProperty("type"), is(equalTo("password")));
+        }
+
+        @TestTemplate
         void shouldReturnInputElementsUnderShadowDom(WebDriver wd) {
             // Given
             pageContent = () -> HTML_SHADOM_DOM;
@@ -1223,6 +1283,49 @@ class AuthUtilsUnitTest extends TestUtils {
             assertThat(inputElements, hasSize(2));
             assertId(inputElements.get(0), "host-input-a");
             assertId(inputElements.get(1), "host-input-b");
+        }
+
+        @TestTemplate
+        void shouldReturnOnFieldOnSubmit(WebDriver wd) {
+            // Given
+            pageContent =
+                    () ->
+                            """
+                                <input type="password" />
+                            """;
+            wd.get(url);
+            WebElement passwordField = wd.findElement(By.tagName("input"));
+            AuthenticationDiagnostics diags = mock();
+
+            // When
+            AuthUtils.submit(diags, wd, passwordField, 1);
+
+            // Then
+            verify(diags).recordStep(wd, "Auto Return");
+            verifyNoMoreInteractions(diags);
+        }
+
+        @TestTemplate
+        void shouldClickButtonIfReturnDoesNoActionOnFieldOnSubmit(WebDriver wd) {
+            // Given
+            pageContent =
+                    () ->
+                            """
+                                <input type="password" />
+                                <button />
+                            """;
+            wd.get(url);
+            WebElement passwordField = wd.findElement(By.tagName("input"));
+            AuthenticationDiagnostics diags = mock();
+
+            // When
+            AuthUtils.submit(diags, wd, passwordField, 1);
+
+            // Then
+            WebElement button = wd.findElement(By.tagName("button"));
+            verify(diags).recordStep(wd, "Auto Return");
+            verify(diags).recordStep(wd, "Click Button", button);
+            verifyNoMoreInteractions(diags);
         }
 
         private static void assertId(WebElement element, String id) {